pyproject.toml
[project]
name = "local-talking-llm"
version = "0.1.0"
description = "A fully local voice assistant — Whisper STT, LLM, and Chatterbox TTS"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-openai>=0.3.0",
    "langchain>=0.3.25",
    "mlx-audio-plus>=0.1.8",
    "openai-whisper>=20240930",
    "nltk>=3.9.1",
    "numpy>=2.1",
    "rich>=14.0.0",
    "sounddevice>=0.5.2",
    "soundfile>=0.13.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "requests>=2.32.0",
    "torchaudio>=2.6.0",
    "ddgs>=7.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
local-talking-llm = "cli:main"
tts-server = "tts_server:main"

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
]

[tool.uv.extra-build-dependencies]
openai-whisper = ["setuptools"]
pkuseg = ["numpy"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["src/"]
sources = ["src"]
